requirements.txt
asgiref==3.9.1
certifi==2025.8.3
dj-database-url==3.0.1
Django==5.2.6
django-jazzmin==3.0.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==23.0.0
packaging==25.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
sib-api-v3-sdk==7.6.0
six==1.17.0
sqlparse==0.5.3
urllib3==2.5.0
whitenoise==6.10.0
